Cleaned up the file uploading code
There was some dead code in there and parts that didn't make sense. For instance
it was unecessary to check if the file was a string or a file handle to when
figuring out the size because by then we had already opened up a file handle
and ditched the string. Further, we weren't closing the file handle which means
file handle leaks.